Map, where
list-of-pairs constructor upgraded in order to not prevent copy and move constructors calls. Now it requires C++17 for "fold" in templates.
"explicit" for move and copy constructors was a mistake so removed.
traversal now supports functors (generally, any objects with operator()).
map and where functions added.
also few small fixes.